HexMTL1, building JSON files in Java, GetRelativePath all done
This commit saw the near completion of Java classes which build the JSON file necessary to send to the front end. The final part of this process is to make a Java class that keeps track of hex coordinates for each riding. Adds them, stores them in a JSON file.